+++ package.json
@@ -10,6 +10,48 @@
       "require": "./dist/index.cjs",
       "node": "./dist/index.cjs",
       "default": "./dist/index.js"
+    },
+    "./dist/aux/auxiliary": {
+      "types": "./dist/aux/auxiliary.d.ts",
+      "import": "./dist/aux/auxiliary.js",
+      "require": "./dist/aux/auxiliary.cjs",
+      "node": "./dist/aux/auxiliary.cjs",
+      "default": "./dist/aux/auxiliary.js"
+    },
+    "./dist/aux/extractor": {
+      "types": "./dist/aux/extractor.d.ts",
+      "import": "./dist/aux/extractor.js",
+      "require": "./dist/aux/extractor.cjs",
+      "node": "./dist/aux/extractor.cjs",
+      "default": "./dist/aux/extractor.js"
+    },
+    "./dist/aux/regexpes": {
+      "types": "./dist/aux/regexpes.d.ts",
+      "import": "./dist/aux/regexpes.js",
+      "require": "./dist/aux/regexpes.cjs",
+      "node": "./dist/aux/regexpes.cjs",
+      "default": "./dist/aux/regexpes.js"
+    },
+    "./dist/aux/tlds": {
+      "types": "./dist/aux/tlds.d.ts",
+      "import": "./dist/aux/tlds.js",
+      "require": "./dist/aux/tlds.cjs",
+      "node": "./dist/aux/tlds.cjs",
+      "default": "./dist/aux/tlds.js"
+    },
+    "./dist/aux/utils": {
+      "types": "./dist/aux/utils.d.ts",
+      "import": "./dist/aux/utils.js",
+      "require": "./dist/aux/utils.cjs",
+      "node": "./dist/aux/utils.cjs",
+      "default": "./dist/aux/utils.js"
+    },
+    "./dist/stix2": {
+      "types": "./dist/sitx2/index.d.ts",
+      "import": "./dist/stix2/index.js",
+      "require": "./dist/sitx2/index.cjs",
+      "node": "./dist/stix2/index.cjs",
+      "default": "./dist/sitx2/index.js"
     }
   },
   "main": "./dist/index.cjs",
@@ -70,7 +112,7 @@
   "devDependencies": {
     "@types/array-unique": "0.3.0",
     "@types/benchmark": "2.1.2",
-    "@types/jest": "29.2.5",
+    "@types/jest": "^29.2.5",
     "@types/memoizee": "0.4.8",
     "@types/node": "18.11.18",
     "@types/punycode": "2.1.0",
@@ -104,3 +146,4 @@
     "uuid": "^9.0.0"
   }
 }
+
