+++ src/stix2/index.ts
